CharacterGamePiece Buttons Implementation
This commit visually does not change anything, but passes in real
CharacterGamePieces into the JButtons instead of using dummy "?" chars.
Now, the program will be able to use the methods of CharacterGamePieces
in the implementation of button pressing. MemoryGameFrame was also
changed to MemoryGameBoard to better follow my UML Diagram.