Partial handle() implementation
This commit adds the first pieces of logic into the KeyboardButtonAction
and MemoryGameModel classes necessary to facilitate a turn. A character
card cannot be selected to flip if it is already flipped - if so, an
instruction label at the top now gives info about the current state of
the game and can warn users about their error. A turn label has also
been added, but as of now just increments every time a card is flipped.