Handle Guess Logic
This commit fills out the handleGuess method in the model so that a
player's actions are correctly reflected on the board. The model now has
a firstGuess and secondGuess variable that it keeps and updates
throughout the game and allows the class to compare the two and handle
that comparison correctly. If the two match, the cards stay flipped over
and "Match!" is displayed. If not, "No match" is displayed, the system
waits two seconds, and then the cards are flipped back over. A "freeze"
has also been added to the program so that no inputs are valid during
those two seconds. This is the first playable version of the game.