Leaderboard Implementation
This commit adds a leaderboard to the game. The leaderboard is shown at
the start of every game, and updated at the end of every game. One
player is displayed along with their high score for each of the
difficulties. The leaderboard data is stored persistently through the
leaderboard.txt file in the resources folder.